Improved karma
Changed the Karma controller to VotesController to align with the vote model. Both create and update actions are performed under the update method, but it cant be helped as only the update method is routed to, and the controller doesnt know until it starts processing the request whether the user has previously voted on the object or not. Also refactored the controller quite a bit, as there was lot of repeating code.